project lagi

+++ app/lib/actions.ts
@@ -4,7 +4,7 @@ import bcrypt from "bcryptjs";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { revalidatePath } from "next/cache";
-
+import { requireAdmin } from "@/app/lib/auth";
 
 
 
@@ -61,9 +61,10 @@ export async function updateBarang(formData: FormData) {
 
 /* DELETE */
 export async function deleteBarang(id: string) {
+  requireAdmin();
+
   await sql`
-    DELETE FROM barang
-    WHERE id = ${id}
+    DELETE FROM barang WHERE id = ${id}
   `;
 
   revalidatePath("/dashboard");
